Add new "deploy" script and necessary package
Commit includes addition of a new command in package.json to enable easier deployment and the 'gh-pages' package to carry out the deployment. Minor refactoring of import statements and a change in the router basename in main.tsx was also done. The deployments are now automatic using GitHub pages. The refactoring and change were done to ensure best practices are followed in code structure. This commit will enable faster and more efficient deployment of application updates.